+++ lib/run_loop/sim_control.rb
@@ -21,6 +21,16 @@ module RunLoop
     end
 
     # @!visibility private
+    # Are we running Xcode 7 or above?
+    #
+    # This is a convenience method.
+    #
+    # @return [Boolean] `true` if the current Xcode version is >= 7.0
+    def xcode_version_gte_7?
+      xctools.xcode_version_gte_7?
+    end
+
+    # @!visibility private
     # Are we running Xcode 6 or above?
     #
     # This is a convenience method.
@@ -69,26 +79,19 @@ module RunLoop
     end
 
     # If it is not already running, launch the simulator for the current version
-    # of Xcode.
+    # of Xcode.  Launches the simulator in the background so it does not
+    # steal focus.
     #
     # @param [Hash] opts Optional controls.
     # @option opts [Float] :post_launch_wait (2.0) How long to sleep after the
     #  simulator has launched.
-    # @option opts [Boolean] :hide_after (false) If true, will attempt to Hide
-    #  the simulator after it is launched.  This is useful `only when testing
-    #  gem features` that require the simulator be launched repeated and you are
-    #  tired of your editor losing focus. :)
     #
     # @todo Consider migrating apple script call to xctools.
     def launch_sim(opts={})
       unless sim_is_running?
-        default_opts = {:post_launch_wait => RunLoop::Environment.sim_post_launch_wait || 2.0,
-                        :hide_after => false}
+        default_opts = {:post_launch_wait => RunLoop::Environment.sim_post_launch_wait || 2.0}
         merged_opts = default_opts.merge(opts)
-        `xcrun open -a "#{sim_app_path}"`
-        if merged_opts[:hide_after]
-          `xcrun /usr/bin/osascript -e 'tell application "System Events" to keystroke "h" using command down'`
-        end
+        `xcrun open -g -a "#{sim_app_path}"`
         sleep(merged_opts[:post_launch_wait]) if merged_opts[:post_launch_wait]
       end
     end
@@ -101,14 +104,9 @@ module RunLoop
     #  simulator has quit.
     # @option opts [Float] :post_launch_wait (2.0) How long to sleep after the
     #  simulator has launched.
-    # @option opts [Boolean] :hide_after (false) If true, will attempt to Hide
-    #  the simulator after it is launched.  This is useful `only when testing
-    #  gem features` that require the simulator be launched repeated and you are
-    #  tired of your editor losing focus. :)
     def relaunch_sim(opts={})
       default_opts = {:post_quit_wait => 1.0,
-                      :post_launch_wait => RunLoop::Environment.sim_post_launch_wait || 2.0,
-                      :hide_after => false}
+                      :post_launch_wait => RunLoop::Environment.sim_post_launch_wait || 2.0}
       merged_opts = default_opts.merge(opts)
       quit_sim(merged_opts)
       launch_sim(merged_opts)
@@ -132,7 +130,13 @@ module RunLoop
       # SimControl.new.quit_sim({:post_quit_wait => 0.5})
 
       processes =
-            ['iPhone Simulator.app', 'iOS Simulator.app',
+            [
+                  # Xcode < 5.1
+                  'iPhone Simulator.app',
+                  # 7.0 < Xcode <= 6.0
+                  'iOS Simulator.app',
+                  # Xcode >= 7.0
+                  'Simulator.app',
 
              # Multiple launchd_sim processes have been causing problems.  This
              # is a first pass at investigating what it would mean to kill the
@@ -190,17 +194,11 @@ module RunLoop
     #  simulator has launched.  Waits longer than normal because we need the
     #  simulator directories to be repopulated. **NOTE:** This option is ignored
     #  in Xcode 6.
-    # @option opts [Boolean] :hide_after (false) If true, will attempt to Hide
-    #  the simulator after it is launched.  This is useful `only when testing
-    #  gem features` that require the simulator be launched repeated and you are
-    #  tired of your editor losing focus. :) **NOTE:** This option is ignored
-    #  in Xcode 6.
     # @option opts [String] :sim_udid (nil) The udid of the simulator to reset.
     #  **NOTE:** This option is ignored in Xcode < 6.
     def reset_sim_content_and_settings(opts={})
       default_opts = {:post_quit_wait => 1.0,
                       :post_launch_wait => RunLoop::Environment.sim_post_launch_wait || 3.0,
-                      :hide_after => false,
                       :sim_udid => nil}
       merged_opts = default_opts.merge(opts)
 
@@ -606,7 +604,8 @@ module RunLoop
     #
     # @return [String, nil] The pid as a String or nil if no process is found.
     def sim_pid
-      `xcrun ps x -o pid,command | grep "#{sim_name}" | grep -v grep`.strip.split(' ').first
+      process_name = "MacOS/#{sim_name}"
+      `xcrun ps x -o pid,command | grep "#{process_name}" | grep -v grep`.strip.split(' ').first
     end
 
     # @!visibility private
@@ -621,7 +620,9 @@ module RunLoop
     #  launching the current simulator.
     def sim_name
       @sim_name ||= lambda {
-        if xcode_version_gte_6?
+        if xcode_version_gte_7?
+          'Simulator'
+        elsif xcode_version_gte_6?
           'iOS Simulator'
         else
           'iPhone Simulator'
@@ -639,7 +640,9 @@ module RunLoop
     def sim_app_path
       @sim_app_path ||= lambda {
         dev_dir = xctools.xcode_developer_dir
-        if xcode_version_gte_6?
+        if xcode_version_gte_7?
+          "#{dev_dir}/Applications/Simulator.app"
+        elsif xcode_version_gte_6?
           "#{dev_dir}/Applications/iOS Simulator.app"
         else
           "#{dev_dir}/Platforms/iPhoneSimulator.platform/Developer/Applications/iPhone Simulator.app"