+++ spec/lib/core_spec.rb
@@ -25,7 +25,7 @@ describe RunLoop::Core do
                   "/Xcode/6.2/Xcode.app/Contents/Applications/Instruments.app/Contents/Resources/templates/Time Profiler.tracetemplate",
             ]
       expect(xctools).to receive(:instruments).with(:templates).and_return(templates)
-      expect { RunLoop::Core.default_tracetemplate(xctools) }.to raise_error
+      expect { RunLoop::Core.default_tracetemplate(xctools) }.to raise_error(RuntimeError)
     end
   end
 
@@ -89,6 +89,15 @@ describe RunLoop::Core do
       xctools = RunLoop::XCTools.new
       expect(xctools).to receive(:xcode_version).at_least(:once).and_return(version)
       expected = 'iPhone 5s (8.4 Simulator)'
+      actual = RunLoop::Core.default_simulator(xctools)
+      expect(actual).to be == expected
+    end
+
+    it "when Xcode 7.0 it returns 'iPhone 5s (9.0 Simulator)'" do
+      version = RunLoop::Version.new('7.0')
+      xctools = RunLoop::XCTools.new
+      expect(xctools).to receive(:xcode_version).at_least(:once).and_return(version)
+      expected = 'iPhone 5s (9.0 Simulator)'
       actual = RunLoop::Core.default_simulator(xctools)
       expect(actual).to be == expected
     end